back/requirements.txt
Flask
ydb[yc]
python-dotenv
flask-cors